fix(result): refactor detail view
