+++ app/src/main/java/com/lagradost/cloudstream3/ui/result/compose/MovieDetailsComposeScreen.kt
@@ -1,769 +1,47 @@
 package com.lagradost.cloudstream3.ui.result.compose
 
 import android.widget.Toast
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
-import androidx.compose.foundation.combinedClickable
-import androidx.compose.foundation.focusable
 import androidx.compose.foundation.gestures.BringIntoViewSpec
 import androidx.compose.foundation.gestures.LocalBringIntoViewSpec
 import androidx.compose.foundation.interaction.MutableInteractionSource
-import androidx.compose.foundation.interaction.collectIsFocusedAsState
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.clipToBounds
 import androidx.compose.ui.focus.FocusRequester
-import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.graphicsLayer
-import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.semantics.Role
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.zIndex
-import coil3.compose.AsyncImage
 import com.lagradost.cloudstream3.ActorData
 import com.lagradost.cloudstream3.R
 import com.lagradost.cloudstream3.SearchResponse
 import com.lagradost.cloudstream3.ui.result.ResultEpisode
 import com.lagradost.cloudstream3.ui.result.ResumeWatchingStatus
-import com.lagradost.cloudstream3.ui.result.compose.components.CircleActionButton
-import com.lagradost.cloudstream3.ui.result.compose.components.HeroPlayButton
-import com.lagradost.cloudstream3.ui.result.compose.components.HeroTrailerButton
-import com.lagradost.cloudstream3.ui.result.compose.components.MaturityRatingBadge
+import com.lagradost.cloudstream3.ui.result.compose.components.AboutSection
+import com.lagradost.cloudstream3.ui.result.compose.components.EpisodeRowItem
+import com.lagradost.cloudstream3.ui.result.compose.components.EpisodesHeaderSection
+import com.lagradost.cloudstream3.ui.result.compose.components.HeroBannerSection
 import com.lagradost.cloudstream3.ui.result.compose.components.MovieCardItem
-import com.lagradost.cloudstream3.ui.result.compose.components.MovieCardSize
 import com.lagradost.cloudstream3.ui.result.compose.components.MovieCardType
-import com.lagradost.cloudstream3.ui.result.compose.components.MovieDetailsMovieCard
-import com.lagradost.cloudstream3.ui.result.compose.components.MovieDetailsTokens
-import com.lagradost.cloudstream3.ui.result.compose.components.SeasonDropdown
-import com.lagradost.cloudstream3.ui.result.compose.components.VideoQualityBadge
+import com.lagradost.cloudstream3.ui.result.compose.components.MovieInfoSynopsisSection
+import com.lagradost.cloudstream3.ui.result.compose.components.RecommendationRowView
 import com.lagradost.cloudstream3.ui.result.compose.theme.MovieDetailsTheme
-import com.lagradost.cloudstream3.ui.result.compose.theme.PrimaryBlack
-import com.lagradost.cloudstream3.ui.result.compose.theme.PrimaryWhite
-import com.lagradost.cloudstream3.ui.result.compose.theme.TransparentBlack60
-import com.lagradost.cloudstream3.ui.result.compose.theme.getRatingScoreColor
-import com.lagradost.cloudstream3.ui.result.getWatchProgress
-
-@Immutable
-data class MovieTrailerData(
-    val title: String,
-    val runtime: String,
-    val rawTrailer: Any? = null
-)
-
-@Immutable
-data class MovieRecommendationRow(
-    val rowIndex: Int,
-    val items: List<MovieCardItem>
-)
-
-private fun getPlayButtonText(
-    context: android.content.Context,
-    resumeStatus: ResumeWatchingStatus?,
-    episodesToDisplay: List<ResultEpisode>,
-    isMovie: Boolean
-): String {
-    if (resumeStatus != null) {
-        val resumeEp = resumeStatus.result
-        val prefix = if (resumeStatus.progress != null) {
-            context.getString(R.string.resume)
-        } else {
-            context.getString(R.string.play_movie_button)
-        }
-        if (resumeStatus.isMovie) {
-            return prefix
-        }
-        val sShort = context.getString(R.string.season_short)
-        val eShort = context.getString(R.string.episode_short)
-        val s = resumeEp.season
-        val e = resumeEp.episode
-        val epCode = if (s != null && s > 0 && e > 0) {
-            "$sShort$s:$eShort$e"
-        } else if (e > 0) {
-            "$eShort$e"
-        } else {
-            ""
-        }
-        return if (epCode.isNotBlank()) "$prefix $epCode" else prefix
-    }
-    val firstEp = episodesToDisplay.firstOrNull()
-    if (firstEp != null && !isMovie) {
-        val sShort = context.getString(R.string.season_short)
-        val eShort = context.getString(R.string.episode_short)
-        val s = firstEp.season
-        val e = firstEp.episode
-        val epCode = if (s != null && s > 0 && e > 0) {
-            "$sShort$s:$eShort$e"
-        } else if (e > 0) {
-            "$eShort$e"
-        } else {
-            ""
-        }
-        val playStr = context.getString(R.string.play_movie_button)
-        return if (epCode.isNotBlank()) "$playStr $epCode" else playStr
-    }
-    return context.getString(R.string.play_movie_button)
-}
-
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-private fun HeroBannerSection(
-    title: String,
-    providerName: String?,
-    backdropUrl: String?,
-    logoUrl: String?,
-    heroHeight: androidx.compose.ui.unit.Dp,
-    playButtonText: String,
-    resumeProgressFraction: Float?,
-    isInWatchList: Boolean,
-    isFavorite: Boolean,
-    hasTrailers: Boolean,
-    playButtonFocusRequester: FocusRequester,
-    playInteractionSource: MutableInteractionSource,
-    inMyListInteractionSource: MutableInteractionSource,
-    likeInteractionSource: MutableInteractionSource,
-    trailerInteractionSource: MutableInteractionSource,
-    searchInteractionSource: MutableInteractionSource,
-    onPlayClick: () -> Unit,
-    onPlayLongClick: (() -> Unit)?,
-    onAddToListClick: () -> Unit,
-    onLikeClick: () -> Unit,
-    onTrailerClick: () -> Unit,
-    onSearchClick: (() -> Unit)?
-) {
-    val colors = MovieDetailsTheme.colors
-    val typography = MovieDetailsTheme.typography
-    val dimens = MovieDetailsTheme.dimens
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .height(heroHeight)
-            .clipToBounds()
-            .background(colors.background)
-    ) {
-        if (!backdropUrl.isNullOrBlank()) {
-            AsyncImage(
-                model = backdropUrl,
-                contentDescription = null,
-                contentScale = ContentScale.Crop,
-                modifier = Modifier
-                    .fillMaxSize()
-                    .clipToBounds()
-            )
-        }
-
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .fillMaxHeight(0.85f)
-                .align(Alignment.BottomCenter)
-                .background(
-                    Brush.verticalGradient(
-                        colors = listOf(
-                            Color.Transparent,
-                            colors.background.copy(alpha = 0.5f),
-                            colors.background.copy(alpha = 0.92f),
-                            colors.background
-                        )
-                    )
-                )
-        )
-
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(Alignment.TopStart)
-                .padding(horizontal = dimens.spacing2Xl, vertical = dimens.spacing2Xl)
-                .zIndex(50f),
-            verticalArrangement = Arrangement.spacedBy(dimens.spacingM)
-        ) {
-            Box(
-                modifier = Modifier
-                    .height(42.dp)
-                    .width(110.dp)
-                    .focusRequester(playButtonFocusRequester)
-                    .combinedClickable(
-                        interactionSource = playInteractionSource,
-                        indication = null,
-                        role = Role.Button,
-                        onClick = onPlayClick,
-                        onLongClick = onPlayLongClick
-                    )
-            )
-
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(dimens.spacingM)
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .size(40.dp)
-                            .clickable(
-                                interactionSource = inMyListInteractionSource,
-                                indication = null,
-                                role = Role.Button,
-                                onClick = onAddToListClick
-                            )
-                    )
-
-                    Box(
-                        modifier = Modifier
-                            .size(40.dp)
-                            .clickable(
-                                interactionSource = likeInteractionSource,
-                                indication = null,
-                                role = Role.Button,
-                                onClick = onLikeClick
-                            )
-                    )
-
-                    if (onSearchClick != null) {
-                        Box(
-                            modifier = Modifier
-                                .size(40.dp)
-                                .clickable(
-                                    interactionSource = searchInteractionSource,
-                                    indication = null,
-                                    role = Role.Button,
-                                    onClick = onSearchClick
-                                )
-                        )
-                    }
-                }
-
-                if (hasTrailers) {
-                    Box(
-                        modifier = Modifier
-                            .height(40.dp)
-                            .width(120.dp)
-                            .clickable(
-                                interactionSource = trailerInteractionSource,
-                                indication = null,
-                                role = Role.Button,
-                                onClick = onTrailerClick
-                            )
-                    )
-                }
-            }
-        }
-
-        Box(
-            modifier = Modifier
-                .align(Alignment.TopEnd)
-                .padding(dimens.spacing2Xl)
-        ) {
-            if (!providerName.isNullOrBlank()) {
-                Text(
-                    text = providerName.uppercase(),
-                    style = typography.boldTitle2.copy(
-                        fontWeight = FontWeight.Black,
-                        letterSpacing = (-0.8).sp
-                    ),
-                    color = colors.primary
-                )
-            }
-        }
-
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .align(Alignment.BottomStart)
-                .padding(horizontal = dimens.spacing2Xl, vertical = dimens.spacingL),
-            horizontalArrangement = Arrangement.SpaceBetween,
-            verticalAlignment = Alignment.Bottom
-        ) {
-            Column {
-                if (!logoUrl.isNullOrBlank()) {
-                    Box(
-                        modifier = Modifier
-                            .height(68.dp)
-                            .width(240.dp)
-                    ) {
-                        AsyncImage(
-                            model = logoUrl,
-                            contentDescription = title,
-                            contentScale = ContentScale.Fit,
-                            modifier = Modifier.fillMaxSize()
-                        )
-                    }
-                } else {
-                    Text(
-                        text = title,
-                        style = typography.boldTitle1,
-                        fontSize = 32.sp,
-                        color = colors.textPrimary,
-                        maxLines = 2,
-                        overflow = TextOverflow.Ellipsis
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(dimens.spacingL))
-
-                HeroPlayButton(
-                    onClick = onPlayClick,
-                    onLongClick = onPlayLongClick,
-                    text = playButtonText,
-                    progress = resumeProgressFraction,
-                    interactionSource = playInteractionSource,
-                    enabled = false
-                )
-
-                Spacer(modifier = Modifier.height(dimens.spacingM))
-
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(dimens.spacingM)
-                ) {
-                    CircleActionButton(
-                        icon = painterResource(
-                            id = if (isInWatchList) R.drawable.ic_baseline_check_24 else R.drawable.ic_baseline_add_24
-                        ),
-                        contentDescription = if (isInWatchList) {
-                            stringResource(id = R.string.in_my_list)
-                        } else {
-                            stringResource(id = R.string.add_to_my_list)
-                        },
-                        onClick = onAddToListClick,
-                        interactionSource = inMyListInteractionSource,
-                        enabled = false
-                    )
-
-                    CircleActionButton(
-                        icon = painterResource(
-                            id = if (isFavorite) R.drawable.ic_baseline_favorite_24 else R.drawable.ic_baseline_favorite_border_24
-                        ),
-                        contentDescription = if (isFavorite) {
-                            stringResource(id = R.string.unfavorite)
-                        } else {
-                            stringResource(id = R.string.favorite)
-                        },
-                        onClick = onLikeClick,
-                        interactionSource = likeInteractionSource,
-                        enabled = false
-                    )
-
-                    if (onSearchClick != null) {
-                        CircleActionButton(
-                            icon = painterResource(id = R.drawable.search_icon),
-                            contentDescription = stringResource(id = R.string.title_search),
-                            onClick = onSearchClick,
-                            interactionSource = searchInteractionSource,
-                            enabled = false
-                        )
-                    }
-                }
-            }
-
-            if (hasTrailers) {
-                HeroTrailerButton(
-                    text = stringResource(id = R.string.play_trailer),
-                    onClick = onTrailerClick,
-                    interactionSource = trailerInteractionSource,
-                    enabled = false
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun MovieInfoSynopsisSection(
-    matchScore: String?,
-    releaseYear: String?,
-    seasonsCount: String?,
-    quality: String?,
-    maturityRating: String?,
-    advisories: String?,
-    top10RankText: String?,
-    synopsis: String,
-    castList: List<String>,
-    genres: List<String>,
-    moodTags: List<String>
-) {
-    val colors = MovieDetailsTheme.colors
-    val typography = MovieDetailsTheme.typography
-    val dimens = MovieDetailsTheme.dimens
-    val hasRightColumn = castList.isNotEmpty() || genres.isNotEmpty() || moodTags.isNotEmpty()
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = dimens.spacing2Xl, vertical = dimens.spacingL),
-        horizontalArrangement = Arrangement.spacedBy(dimens.spacing2Xl)
-    ) {
-        Column(
-            modifier = Modifier
-                .weight(if (hasRightColumn) 0.65f else 1f)
-                .fillMaxWidth(),
-            verticalArrangement = Arrangement.spacedBy(dimens.spacingM)
-        ) {
-            val hasLine1 = !matchScore.isNullOrBlank() || !releaseYear.isNullOrBlank() ||
-                    !seasonsCount.isNullOrBlank() || !quality.isNullOrBlank()
-
-            if (hasLine1) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(dimens.spacingM)
-                ) {
-                    if (!matchScore.isNullOrBlank()) {
-                        Text(
-                            text = matchScore,
-                            color = getRatingScoreColor(matchScore),
-                            style = typography.mediumBody,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-
-                    if (!releaseYear.isNullOrBlank()) {
-                        Text(
-                            text = releaseYear,
-                            color = colors.textPrimary,
-                            style = typography.regularBody
-                        )
-                    }
-
-                    if (!seasonsCount.isNullOrBlank()) {
-                        Text(
-                            text = seasonsCount,
-                            color = colors.textPrimary,
-                            style = typography.regularBody
-                        )
-                    }
-
-                    if (!quality.isNullOrBlank()) {
-                        VideoQualityBadge(quality = quality)
-                    }
-                }
-            }
-
-            if (!maturityRating.isNullOrBlank() || !advisories.isNullOrBlank()) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(dimens.spacingS)
-                ) {
-                    if (!maturityRating.isNullOrBlank()) {
-                        MaturityRatingBadge(rating = maturityRating)
-                    }
-                    if (!advisories.isNullOrBlank()) {
-                        Text(
-                            text = advisories,
-                            color = colors.textMuted,
-                            style = typography.regularCaption1
-                        )
-                    }
-                }
-            }
-
-            if (!top10RankText.isNullOrBlank()) {
-                Row(
-                    verticalAlignment = Alignment.CenterVertically,
-                    horizontalArrangement = Arrangement.spacedBy(dimens.spacingS),
-                    modifier = Modifier.padding(vertical = dimens.spacingXs)
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .clip(MovieDetailsTokens.ShapeCardSmall)
-                            .background(colors.primary)
-                            .padding(horizontal = 6.dp, vertical = 2.dp)
-                    ) {
-                        Text(
-                            text = stringResource(id = R.string.top_10_badge_text),
-                            style = typography.regularCaption2,
-                            color = colors.onPrimary,
-                            fontWeight = FontWeight.Black
-                        )
-                    }
-                    Text(
-                        text = top10RankText,
-                        style = typography.mediumBody,
-                        fontWeight = FontWeight.Bold,
-                        color = colors.textPrimary
-                    )
-                }
-            }
-
-            if (synopsis.isNotBlank()) {
-                val synopsisInteractionSource = remember { MutableInteractionSource() }
-                val isSynopsisFocused by synopsisInteractionSource.collectIsFocusedAsState()
-
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clip(MovieDetailsTokens.ShapeCardSmall)
-                        .background(if (isSynopsisFocused) colors.surfaceElevated.copy(alpha = 0.5f) else Color.Transparent)
-                        .then(
-                            if (isSynopsisFocused) {
-                                Modifier.border(
-                                    BorderStroke(dimens.borderFocus, colors.primary),
-                                    MovieDetailsTokens.ShapeCardSmall
-                                )
-                            } else Modifier
-                        )
-                        .focusable(interactionSource = synopsisInteractionSource)
-                        .padding(dimens.spacingS)
-                ) {
-                    Text(
-                        text = synopsis,
-                        style = typography.regularBody,
-                        color = if (isSynopsisFocused) colors.textPrimary else colors.textSecondary,
-                        lineHeight = 22.sp
-                    )
-                }
-            }
-        }
-
-        if (hasRightColumn) {
-            Column(
-                modifier = Modifier
-                    .weight(0.35f)
-                    .fillMaxWidth(),
-                verticalArrangement = Arrangement.spacedBy(dimens.spacingM)
-            ) {
-                if (castList.isNotEmpty()) {
-                    Column(verticalArrangement = Arrangement.spacedBy(2.dp)) {
-                        Text(
-                            text = stringResource(id = R.string.cast_label),
-                            style = typography.regularCaption2,
-                            color = colors.textSecondary
-                        )
-                        Text(
-                            text = castList.joinToString(", "),
-                            style = typography.regularCaption1,
-                            color = colors.textPrimary,
-                            maxLines = 3,
-                            overflow = TextOverflow.Ellipsis
-                        )
-                    }
-                }
-
-                if (genres.isNotEmpty()) {
-                    Column(verticalArrangement = Arrangement.spacedBy(2.dp)) {
-                        Text(
-                            text = stringResource(id = R.string.genres_label),
-                            style = typography.regularCaption2,
-                            color = colors.textSecondary
-                        )
-                        Text(
-                            text = genres.joinToString(", "),
-                            style = typography.regularCaption1,
-                            color = colors.textPrimary
-                        )
-                    }
-                }
-
-                if (moodTags.isNotEmpty()) {
-                    Column(verticalArrangement = Arrangement.spacedBy(2.dp)) {
-                        Text(
-                            text = stringResource(id = R.string.mood_tags_label),
-                            style = typography.regularCaption2,
-                            color = colors.textSecondary
-                        )
-                        Text(
-                            text = moodTags.joinToString(", "),
-                            style = typography.regularCaption1,
-                            color = colors.textPrimary
-                        )
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun EpisodesHeaderSection(
-    seasonOptions: List<String>,
-    selectedSeasonText: String,
-    onSeasonSelect: ((Int) -> Unit)?,
-    onSeasonTextChange: (String) -> Unit,
-    dubOptions: List<String>,
-    selectedDubText: String,
-    onDubSelect: ((Int) -> Unit)?,
-    onDubTextChange: (String) -> Unit,
-    rangeOptions: List<String>,
-    selectedRangeText: String,
-    onRangeSelect: ((Int) -> Unit)?,
-    onRangeTextChange: (String) -> Unit
-) {
-    val colors = MovieDetailsTheme.colors
-    val typography = MovieDetailsTheme.typography
-    val dimens = MovieDetailsTheme.dimens
-
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = dimens.spacing2Xl)
-            .padding(top = dimens.spacing2Xl, bottom = dimens.spacingL),
-        horizontalArrangement = Arrangement.spacedBy(dimens.spacingL),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Text(
-            text = stringResource(id = R.string.episodes),
-            style = typography.boldTitle2,
-            fontSize = 24.sp,
-            color = colors.textPrimary
-        )
-
-        if (seasonOptions.size > 1) {
-            SeasonDropdown(
-                options = seasonOptions,
-                selectedOption = selectedSeasonText,
-                onOptionSelected = { seasonStr ->
-                    onSeasonTextChange(seasonStr)
-                    val idx = seasonOptions.indexOf(seasonStr)
-                    if (idx >= 0) onSeasonSelect?.invoke(idx)
-                },
-                width = 180.dp
-            )
-        }
-
-        if (dubOptions.size > 1) {
-            SeasonDropdown(
-                options = dubOptions,
-                selectedOption = selectedDubText,
-                onOptionSelected = { dubStr ->
-                    onDubTextChange(dubStr)
-                    val idx = dubOptions.indexOf(dubStr)
-                    if (idx >= 0) onDubSelect?.invoke(idx)
-                },
-                width = 140.dp
-            )
-        }
-
-        if (rangeOptions.size > 1) {
-            SeasonDropdown(
-                options = rangeOptions,
-                selectedOption = selectedRangeText,
-                onOptionSelected = { rangeStr ->
-                    onRangeTextChange(rangeStr)
-                    val idx = rangeOptions.indexOf(rangeStr)
-                    if (idx >= 0) onRangeSelect?.invoke(idx)
-                },
-                width = 140.dp
-            )
-        }
-    }
-}
-
-@Composable
-private fun AboutSection(
-    title: String,
-    creator: String?,
-    castList: List<String>,
-    writers: List<String>,
-    genres: List<String>,
-    moodTags: List<String>,
-    maturityRating: String?,
-    advisories: String?
-) {
-    val colors = MovieDetailsTheme.colors
-    val typography = MovieDetailsTheme.typography
-    val dimens = MovieDetailsTheme.dimens
-    val aboutInteractionSource = remember { MutableInteractionSource() }
-    val isAboutFocused by aboutInteractionSource.collectIsFocusedAsState()
-    val aboutBorder = if (isAboutFocused) BorderStroke(dimens.borderFocus, colors.primary) else BorderStroke(1.dp, colors.border)
-
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = dimens.spacing2Xl)
-            .padding(top = dimens.spacing3Xl, bottom = dimens.spacing3Xl)
-    ) {
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clip(MovieDetailsTokens.ShapeCardMedium)
-                .background(if (isAboutFocused) colors.surfaceElevated else colors.surface)
-                .border(aboutBorder, MovieDetailsTokens.ShapeCardMedium)
-                .focusable(interactionSource = aboutInteractionSource)
-                .padding(dimens.spacing2Xl),
-            verticalArrangement = Arrangement.spacedBy(dimens.spacingM)
-        ) {
-            Text(
-                text = stringResource(id = R.string.about_title_format, title),
-                style = typography.boldTitle2,
-                fontSize = 20.sp,
-                color = colors.textPrimary
-            )
-
-            Spacer(modifier = Modifier.height(dimens.spacingXs))
-
-            if (!creator.isNullOrBlank()) {
-                AboutMetadataRow(label = stringResource(id = R.string.creator_label), value = creator)
-            }
-            if (castList.isNotEmpty()) {
-                AboutMetadataRow(label = stringResource(id = R.string.cast_label), value = castList.joinToString(", "))
-            }
-            if (writers.isNotEmpty()) {
-                AboutMetadataRow(label = stringResource(id = R.string.writers_label), value = writers.joinToString(", "))
-            }
-            if (genres.isNotEmpty()) {
-                AboutMetadataRow(label = stringResource(id = R.string.genres_label), value = genres.joinToString(", "))
-            }
-            if (moodTags.isNotEmpty()) {
-                AboutMetadataRow(label = stringResource(id = R.string.mood_tags_label), value = moodTags.joinToString(", "))
-            }
-            if (!maturityRating.isNullOrBlank()) {
-                val ratingDescription = if (!advisories.isNullOrBlank()) {
-                    "$maturityRating ($advisories)"
-                } else {
-                    maturityRating
-                }
-                AboutMetadataRow(
-                    label = stringResource(id = R.string.maturity_rating_label),
-                    value = ratingDescription
-                )
-            }
-        }
-    }
-}
 
 @OptIn(ExperimentalFoundationApi::class)
 @Composable
@@ -1085,304 +363,6 @@ fun MovieDetailsComposeScreen(
     }
 }
 
-@OptIn(ExperimentalFoundationApi::class)
-@Composable
-fun EpisodeRowItem(
-    episode: ResultEpisode,
-    onClick: () -> Unit,
-    onLongClick: (() -> Unit)? = null,
-    modifier: Modifier = Modifier
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    val isFocused by interactionSource.collectIsFocusedAsState()
-    val typography = MovieDetailsTheme.typography
-    val dimens = MovieDetailsTheme.dimens
-    val colors = MovieDetailsTheme.colors
-
-    val scaleState = animateFloatAsState(
-        targetValue = if (isFocused) 1.02f else 1f,
-        animationSpec = MovieDetailsTokens.FastFocusAnimationSpec,
-        label = "episodeScale"
-    )
-
-    val background = if (isFocused) colors.surfaceElevated else colors.surface
-    val border = if (isFocused) BorderStroke(dimens.borderFocus, colors.primary) else null
-    val progress = remember(episode) { episode.getWatchProgress() }
-
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .zIndex(if (isFocused) 10f else 0f)
-            .graphicsLayer {
-                scaleX = scaleState.value
-                scaleY = scaleState.value
-            }
-            .clip(MovieDetailsTokens.ShapeCardMedium)
-            .background(background)
-            .then(if (border != null) Modifier.border(border, MovieDetailsTokens.ShapeCardMedium) else Modifier)
-            .combinedClickable(
-                interactionSource = interactionSource,
-                indication = null,
-                role = Role.Button,
-                onClick = onClick,
-                onLongClick = onLongClick
-            )
-            .focusable(interactionSource = interactionSource)
-            .padding(dimens.spacingL),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(dimens.spacingL)
-    ) {
-        Text(
-            text = episode.episode.toString(),
-            style = typography.boldTitle1,
-            fontSize = 24.sp,
-            color = colors.textSecondary,
-            modifier = Modifier.width(32.dp)
-        )
-
-        Box(
-            modifier = Modifier
-                .width(160.dp)
-                .height(90.dp)
-                .clip(MovieDetailsTokens.ShapeCardSmall)
-                .background(colors.surface)
-        ) {
-            if (!episode.poster.isNullOrBlank()) {
-                AsyncImage(
-                    model = episode.poster,
-                    contentDescription = episode.name,
-                    contentScale = ContentScale.Crop,
-                    modifier = Modifier.fillMaxSize()
-                )
-            }
-
-            Box(
-                modifier = Modifier
-                    .size(36.dp)
-                    .clip(CircleShape)
-                    .background(TransparentBlack60)
-                    .border(BorderStroke(1.dp, PrimaryWhite), CircleShape)
-                    .align(Alignment.Center)
-            ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_baseline_play_arrow_24),
-                    contentDescription = null,
-                    tint = PrimaryWhite,
-                    modifier = Modifier
-                        .size(16.dp)
-                        .align(Alignment.Center)
-                )
-            }
-
-            if (progress > 0.05f) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .height(4.dp)
-                        .align(Alignment.BottomStart)
-                        .background(colors.border)
-                ) {
-                    Box(
-                        modifier = Modifier
-                            .fillMaxWidth(progress.coerceIn(0f, 1f))
-                            .fillMaxHeight()
-                            .background(colors.primary)
-                    )
-                }
-            }
-        }
-
-        Column(
-            modifier = Modifier
-                .weight(1f)
-                .fillMaxWidth(),
-            verticalArrangement = Arrangement.spacedBy(4.dp)
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceBetween,
-                verticalAlignment = Alignment.CenterVertically
-            ) {
-                Text(
-                    text = "${episode.episode}. ${episode.name ?: episode.headerName}",
-                    style = typography.mediumBody,
-                    fontWeight = FontWeight.Bold,
-                    color = colors.textPrimary,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis
-                )
-                Text(
-                    text = if (episode.runTime != null) "${episode.runTime}m" else "",
-                    style = typography.regularCaption1,
-                    color = colors.textSecondary
-                )
-            }
-
-            if (!episode.description.isNullOrBlank()) {
-                Text(
-                    text = episode.description,
-                    style = typography.regularCaption1,
-                    color = colors.textMuted,
-                    maxLines = 2,
-                    overflow = TextOverflow.Ellipsis,
-                    lineHeight = 18.sp
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun TrailerItemCard(
-    trailer: MovieTrailerData,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val interactionSource = remember { MutableInteractionSource() }
-    val isFocused by interactionSource.collectIsFocusedAsState()
-    val typography = MovieDetailsTheme.typography
-    val dimens = MovieDetailsTheme.dimens
-    val colors = MovieDetailsTheme.colors
-
-    val scaleState = animateFloatAsState(
-        targetValue = if (isFocused) MovieDetailsTokens.FOCUS_SCALE_FACTOR else 1f,
-        animationSpec = MovieDetailsTokens.FastFocusAnimationSpec,
-        label = "trailerScale"
-    )
-
-    val border = if (isFocused) BorderStroke(dimens.borderFocus, colors.primary) else BorderStroke(dimens.borderSubtle, colors.border)
-
-    Column(
-        modifier = modifier
-            .fillMaxWidth()
-            .zIndex(if (isFocused) 10f else 0f)
-            .graphicsLayer {
-                scaleX = scaleState.value
-                scaleY = scaleState.value
-            }
-            .clip(MovieDetailsTokens.ShapeCardMedium)
-            .background(colors.surfaceElevated)
-            .border(border, MovieDetailsTokens.ShapeCardMedium)
-            .clickable(
-                interactionSource = interactionSource,
-                indication = null,
-                role = Role.Button,
-                onClick = onClick
-            )
-            .focusable(interactionSource = interactionSource)
-    ) {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .height(130.dp)
-                .background(colors.surface)
-        ) {
-            Box(
-                modifier = Modifier
-                    .size(42.dp)
-                    .clip(CircleShape)
-                    .background(TransparentBlack60)
-                    .border(BorderStroke(1.5.dp, PrimaryWhite), CircleShape)
-                    .align(Alignment.Center)
-            ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.ic_baseline_play_arrow_24),
-                    contentDescription = null,
-                    tint = PrimaryWhite,
-                    modifier = Modifier
-                        .size(20.dp)
-                        .align(Alignment.Center)
-                )
-            }
-        }
-
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(dimens.spacingM),
-            verticalArrangement = Arrangement.spacedBy(4.dp)
-        ) {
-            Text(
-                text = trailer.title,
-                style = typography.mediumBody,
-                fontWeight = FontWeight.Bold,
-                color = colors.textPrimary,
-                maxLines = 1,
-                overflow = TextOverflow.Ellipsis
-            )
-            Text(
-                text = trailer.runtime,
-                style = typography.regularCaption2,
-                color = colors.textSecondary
-            )
-        }
-    }
-}
-
-@Composable
-fun AboutMetadataRow(
-    label: String,
-    value: String,
-    modifier: Modifier = Modifier
-) {
-    val typography = MovieDetailsTheme.typography
-    val colors = MovieDetailsTheme.colors
-
-    Row(
-        modifier = modifier.fillMaxWidth(),
-        horizontalArrangement = Arrangement.spacedBy(8.dp)
-    ) {
-        Text(
-            text = if (label.endsWith(":")) label else "$label:",
-            style = typography.regularCaption1,
-            color = colors.textSecondary,
-            modifier = Modifier.width(130.dp)
-        )
-        Text(
-            text = value,
-            style = typography.regularCaption1,
-            color = colors.textPrimary,
-            modifier = Modifier.weight(1f)
-        )
-    }
-}
-
-@Composable
-fun RecommendationRowView(
-    row: MovieRecommendationRow,
-    onCardClick: (MovieCardItem) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val dimens = MovieDetailsTheme.dimens
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = dimens.spacing2Xl, vertical = 6.dp),
-        horizontalArrangement = Arrangement.spacedBy(dimens.spacingM)
-    ) {
-        row.items.forEach { cardItem ->
-            Box(modifier = Modifier.weight(1f)) {
-                MovieDetailsMovieCard(
-                    title = cardItem.title,
-                    type = MovieCardType.POSTER,
-                    size = MovieCardSize.MEDIUM,
-                    badge = cardItem.badge,
-                    posterUrl = cardItem.posterUrl,
-                    showLogo = false,
-                    showBottomTitle = true,
-                    onClick = { onCardClick(cardItem) },
-                    modifier = Modifier.fillMaxWidth()
-                )
-            }
-        }
-        if (row.items.size < 6) {
-            repeat(6 - row.items.size) {
-                Spacer(modifier = Modifier.weight(1f))
-            }
-        }
-    }
-}
-
 @Preview(
     name = "TV 1080p Landscape",
     device = "spec:width=1920dp,height=1080dp,dpi=320,orientation=landscape",
@@ -1393,23 +373,21 @@ fun RecommendationRowView(
 private fun MovieDetailsComposeScreenPreview() {
     MovieDetailsTheme {
         MovieDetailsComposeScreen(
-            title = "HOUSE OF NINJAS",
-            providerName = "SUPERSTREAM",
+            title = "Stranger Things",
+            providerName = "Netflix",
+            backdropUrl = null,
             matchScore = "98% Match",
             releaseYear = "2024",
-            seasonsCount = "1 Season",
-            quality = "4K Ultra HD",
-            maturityRating = "TV-MA",
-            advisories = "smoking, violence, language",
-            top10RankText = "#2 in TV Shows Today",
-            synopsis = "Years after retiring from their formidable ninja lives, a dysfunctional family must return to shadowy missions to counteract a string of looming threats.",
-            cast = listOf("Kento Kaku", "Yosuke Eguchi", "Tae Kimura", "Kengo Kora"),
-            genres = listOf("TV Dramas", "Action", "Japanese", "Suspenseful"),
-            moodTags = listOf("Dark", "Suspenseful", "Exciting"),
-            creator = "Dave Boyle",
-            writers = listOf("Dave Boyle", "Masahiro Yamaura", "Kota Oishi"),
-            isInWatchList = false,
-            hasTrailers = true
+            seasonsCount = "4 Seasons",
+            quality = "4K ULTRA HD",
+            maturityRating = "16+",
+            advisories = "fear, language, violence",
+            top10RankText = "#1 in TV Shows Today",
+            synopsis = "When a young boy vanishes, a small town uncovers a mystery involving secret experiments, terrifying supernatural forces and one strange little girl.",
+            cast = listOf("Winona Ryder", "David Harbour", "Millie Bobby Brown", "Finn Wolfhard"),
+            genres = listOf("Sci-Fi", "Horror", "Drama"),
+            moodTags = listOf("Ominous", "Nostalgic", "Suspenseful"),
+            creator = "The Duffer Brothers"
         )
     }
 }
